Refactoring DLL, added functions for creating server sockets and accepting connections.
